fix: missing test utils
